fixed custom adder 0 amount updates, dialog box emitter issues, and cancel behavior

+++ src/main/java/com/albatross/api/v1/company/blueraven/models/ProposalAdderRequest.java
@@ -2,6 +2,8 @@ package com.albatross.api.v1.company.blueraven.models;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -13,7 +15,7 @@ public class ProposalAdderRequest {
   @NoArgsConstructor
   public static class AdderItem {
     private Long id;
-    private Long customAdderAmount;
+    private BigDecimal customAdderAmount;
     private String adderType;
     private String fieldName;
     private Boolean applied = true;
